Fix 10-blur_portion: accumulate in float, not double
The reference blur accumulates the weighted sums and the weight in 32-bit
float. Using double produced off-by-one pixels (e.g. a uniform 255 region
gives 254.9997 in float -> 254, but exactly 255 in double), so every
output.pbm differed from the checker's reference. Verified byte-identical
against the grader's image_0/kernel_0 outputs for the middle-portion case
(1/2/10 rounds) and the full-image case (borders included).

+++ multithreading/10-blur_portion.c
@@ -17,7 +17,7 @@ static void blur_pixel(blur_portion_t const *portion, size_t px, size_t py)
 	kernel_t const *kernel = portion->kernel;
 	size_t kx, ky, half = kernel->size / 2;
 	long ix, iy;
-	double r = 0, g = 0, b = 0, weight = 0, k;
+	float r = 0, g = 0, b = 0, weight = 0, k;
 	pixel_t *dst;
 
 	for (ky = 0; ky < kernel->size; ky++)
